Refactor[]: 분리 notification
서비스에 있던 notification 관련 코드들을 따로 분리함
