+++ app/src/main/java/campus/tech/kakao/map/view/SplashScreen.kt
@@ -130,7 +130,4 @@ class SplashScreen : AppCompatActivity() {
 			}
 		)
 	}
-
-
-
 }
